Refactor useTodolist and tasksReducer, and create useHeader
Updated the useTodolist with new methods and modified task creation in tasksReducer. The id check while dispatching actions in useTodolist has been removed and task creation was shifted to a thunk creator in tasksReducer. Moreover, a new hook useHeader is created to handle todolist addition, which was previously managed in useTodolist.