Implement AppThunkType in tasks and todolists reducers
This commit implements AppThunkType in the tasks and todolists reducers to streamline the action typing. In addition, the corresponding modifications are made in the store.ts file to include the ThunkAction. This makes the code more consistent and manageable by eliminating the need to explicitly declare dispatch types in each thunk action.