Implement status tracking feature in the app
Added a new appReducer to manage the application status across different components. This status is used in the Header component to show a progress bar when loading data. Updated the todolistsReducer, tasksReducer, and store modules to utilize the new reducer and action creators. Additionally, revised tests due to the changes.