Refactor resultCode and add ErrorSnackbar component
The resultCode type in todolists-api.ts is refactored to an enum for cleaner and more type-safe code. Introduced ErrorSnackbar component to handle errors and also included handling error state in the ReduxStoreProviderDecorator. Task creation and updates now have enhanced error handling in the tasksReducer. The error handling leverages the new RESULT_CODE enum and the ErrorSnackbar component to inform users about errors.