Implement user authentication checks and redirects
The commit implementing authentication checks in the application has been integrated. If a user is not logged in but tries to access pages that require authentication, they are now redirected to the login page. Conversely, if a user is already authenticated and tries to access the login page, they are redirected to the main page. A loading component has been added to show progress whilst app initialization. A 'me' endpoint is also added to the auth api to fetch user info on app start.