Refactor code and update references for tasks features
The restructuring of tasks-related files under the 'features/todolists' directory has mandated updates to import paths. In line with this, the code has been refactored with an emphasis on simplification and enhanced readability. Some of the modifications include renaming and re-routing of declarations, removal of unneeded code, and replacement of code with equivalent logic.