Refactor code with Redux Toolkit and rename auth-reducer
The codebase has been refactored to use Redux Toolkit, resulting in clearer and more concise code. The 'auth-reducer' has been renamed to 'auth-slice'. The use of Redux Toolkit simplifies several areas of the application, including the store configuration and the construction of thunks and selectors. This change makes the code more understandable and manageable.